More work on cpu bus changes
need a bus to handle disconguous memory and devices
using memory mapping.  Also a bus to handle interupts
Possible making it on a seperate thread to better simulate
and active device.
The assumption for interupts is that just  interupts are
only handled at the begining of a instruction cycle for
both NMI and INT. So the bus will set a call into the CPU
to set a flag which the cpu main thread can check before
the next fetch cycle.  If raised start to execut the Interupt
routine after pushing items to stack.